remove old json.rs, rename json_new -> json
