+++ tests/json/blocks/image.rs
