+++ tests/json/compose/confirm.rs
