+++ tests/json/compose/text.rs
