+++ tests/json/elems/checkboxes.rs
