+++ tests/json/elems/image.rs
