--- tests/json_new.rs
@@ -1,2 +1,0 @@
-#[path = "json_new/mod.rs"]
-mod json_new;
